pyproject.toml
[tool.poetry]
name = "youtube-api"
version = "0.1.0"
description = ""
authors = ["Glenn Dalida <dalidaglenno6452@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
google-auth = "^2.0.0"
google-api-python-client = "^2.0.0"
google-auth-oauthlib = "^1.2.1"
pandas = "^2.2.2"
openpyxl = "^3.1.5"
isodate = "^0.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
